Refactor to use the new transport/framer system.
This is a huge patch, but there didn't seem to be a good way around it.
It's basically just converting the existing methods to use the new methods,
replacing the hand-rolled futures thing with blackbird promises, removing
cruft used by the old connection class, and adding one new generic to fix
issues with out-of-order processing.