Rename rename SEND-FRAME, remove convenience wrappers, add connection-oriented wrappers.
In doing the encrypted rpc extension prototype, it was getting pretty
annoying to fetch MTGNET-SYS:CONNECTION-FRAMER/MTGNET-SYS:CONNECTION-TRANSPORT
every time I wanted to send a frame, which is naturally a connection-oriented
activity. Since using a connection-oriented method is preferrable, there's
not much reason to have the convenience &REST variants in the lower-level
versions.