Ignore socket errors when disconnecting.
If e.g. the remote end has closed the connection, disconnecting the
socket can cause an error, which will prevent other parts of the
shutdown from completing normally. Instead, make a best-effort to close
the socket and ignore any errors in the actual shutdown. A warning is
signalled for debugging purposes.