Propagate errors when there is no pending op.
This is useful for disconnects, which can happen even when there is no
registered read or write op. Rather than dropping the error, it is
stored and propagated the next time an op is registered (similar to the
way synchronous streams work).

The webapp now seems to handle having the backend disconnect in a
graceful fashion (you get an error, but the app doesn't go into
limbo-mode and will reconnect).