Use strings for JSON object keys.
This allows us to avoid the details of mapping keys to lisp keywords,
which may be problematic if services have different requirements with
regard to case. This also uses explicit field strings for fields that
are part of protocol objects, so we don't have to worry about callers
setting key-conversion options.